refactor: move tests to separate files
